little updates

+++ lab4-postgresql/ViewModels/CategoryViewModel.cs
@@ -2,6 +2,7 @@
 using lab4_postgresql.Views;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.ObjectModel;
+using System.Windows;
 using System.Windows.Input;
 
 namespace lab4_postgresql.ViewModels
@@ -46,6 +47,11 @@ namespace lab4_postgresql.ViewModels
             CategoryWindow win = new CategoryWindow(new Category());
             if (win.ShowDialog() == true)
             {
+                if (win.Category.CategoryName == null)
+                {
+                    ShowErrorMessage();
+                    return;
+                }
                 Category category = win.Category;
                 db.Categories.Add(category);
                 await db.SaveChangesAsync();
@@ -64,6 +70,11 @@ namespace lab4_postgresql.ViewModels
             CategoryWindow win = new CategoryWindow(category);
             if (win.ShowDialog() == true)
             {
+                if (win.Category.CategoryName == null)
+                {
+                    ShowErrorMessage();
+                    return;
+                }
                 selectedCategory.CategoryName = win.Category.CategoryName;
                 db.Entry(selectedCategory).State = EntityState.Modified;
                 await db.SaveChangesAsync();
@@ -72,9 +83,49 @@ namespace lab4_postgresql.ViewModels
         private async void removeCategoryEntity()
         {
             Category? category = selectedCategory;
-            db.Categories.Remove(category);
+            var products = (from p in db.Products.ToList() where p.Category == category select p).ToList();
+            if (products.Count > 0)
+            {
+                string messageBoxText = "Имеются продукты этой категории. Перед удалением категории удалите все продукты, которые ей принадлежат";
+
+                string caption = "Ошибка";
+
+                MessageBoxImage icon = MessageBoxImage.Error;
+
+                MessageBox.Show(messageBoxText, caption, MessageBoxButton.OK, icon);
+
+                return;
+            }
+
+            if (ShowConfirmationWindow() == MessageBoxResult.Yes)
+            {
+                db.Categories.Remove(category);
+            }
+
             await db.SaveChangesAsync();
         }
         private bool CanExecute() => selectedCategory != null;
+
+        private void ShowErrorMessage()
+        {
+            string messageBoxText = "Некорректные значения";
+
+            string caption = "Ошибка";
+
+            MessageBoxImage icon = MessageBoxImage.Error;
+
+            MessageBox.Show(messageBoxText, caption, MessageBoxButton.OK, icon);
+        }
+
+        private MessageBoxResult ShowConfirmationWindow()
+        {
+            string messageBoxText = "Вы уверены? Действие нельзя будет отменить";
+
+            string caption = "Подтвердите действие";
+
+            MessageBoxImage icon = MessageBoxImage.Question;
+
+            return MessageBox.Show(messageBoxText, caption, MessageBoxButton.YesNo, icon);
+        }
     }
 }
